+++ ComputerGraphics_TomogramVisualizer.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32228.430
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ComputerGraphics_TomogramVisualizer", "ComputerGraphics_TomogramVisualizer\ComputerGraphics_TomogramVisualizer.csproj", "{2E9015A7-1D6E-4581-B2E8-CB219FA1E2E6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tomogram_Utilities", "Tomogram_Utilities\Tomogram_Utilities.csproj", "{5EBA3586-0D73-45DF-9F1B-919951D959F5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tomogram_3D", "Tomogram_3D\Tomogram_3D.csproj", "{6F17D282-F530-471E-AD46-E9133B8ACD9B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ComputerGraphics_TomogramVisualizer", "ComputerGraphics_TomogramVisualizer\ComputerGraphics_TomogramVisualizer.csproj", "{2E9015A7-1D6E-4581-B2E8-CB219FA1E2E6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,14 +15,6 @@ Global
 		{2E9015A7-1D6E-4581-B2E8-CB219FA1E2E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2E9015A7-1D6E-4581-B2E8-CB219FA1E2E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2E9015A7-1D6E-4581-B2E8-CB219FA1E2E6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5EBA3586-0D73-45DF-9F1B-919951D959F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5EBA3586-0D73-45DF-9F1B-919951D959F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5EBA3586-0D73-45DF-9F1B-919951D959F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5EBA3586-0D73-45DF-9F1B-919951D959F5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6F17D282-F530-471E-AD46-E9133B8ACD9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F17D282-F530-471E-AD46-E9133B8ACD9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F17D282-F530-471E-AD46-E9133B8ACD9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6F17D282-F530-471E-AD46-E9133B8ACD9B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
