+++ ComputerGraphics_TomogramVisualizer/Bin.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-namespace Tomogram_Utilities
+namespace ComputerGraphics_TomogramVisualizer
 {
     public static class Bin
     {
@@ -8,7 +8,7 @@ namespace Tomogram_Utilities
         public static int X { get; private set; }
         public static int Y { get; private set; }
         public static int Z { get; private set; }
-        public static short[]? array { get; private set; }
+        public static short[]? array;
 
         public static void readBin(string path)
         {
