+++ ComputerGraphics_TomogramVisualizer/ComputerGraphics_TomogramVisualizer.csproj
@@ -13,8 +13,4 @@
     <PackageReference Include="OpenTK.WinForms" Version="4.0.0-pre.6" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Tomogram_Utilities\Tomogram_Utilities.csproj" />
-  </ItemGroup>
-
 </Project>
