--- README.md
@@ -1,25 +1,0 @@
-## ComputerGraphics Tomogram Visualizer
-
-Лабораторная работа по курсу компьютерной графики. Визуализация томограммы.
-
-Использован язык С#, технология OpenGL (библиотека OpenTK)
-
-Работу выполнили: [Damlrca](https://github.com/Damlrca), [Mortus19](https://github.com/Mortus19), [aartyomm](https://github.com/aartyomm).
-
-![screenshot](./screenshots/screenshot.png)
-
-### Особенности работы
-
-- Три варианта визуализации: Quads, QuadStrip, Texture
-- Режимы растягивания томограммы: Stretch (Растягивание), Zoom (Масштабирование)
-- Выбор диапазона для TransferFunction
-
-## Tomogram 3D
-
-3D визуализация томограммы:
-
-- Варианты проекции: перспективная, ортографическая
-- Выбор минимального порога отрисовываемой прозрачности
-- Настройка количества отрисовываемых двумерных слоёв
-
-![screenshot 3D](./screenshots/screenshot_3D.png)
