--- Tomogram_3D/Tomogram_3D.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net5.0-windows</TargetFramework>
-    <Nullable>enable</Nullable>
-    <UseWindowsForms>true</UseWindowsForms>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="OpenTK" Version="4.7.1" />
-    <PackageReference Include="OpenTK.WinForms" Version="4.0.0-pre.6" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Tomogram_Utilities\Tomogram_Utilities.csproj" />
-  </ItemGroup>
-
-</Project>
