--- Tomogram_Utilities/Tomogram_Utilities.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows</TargetFramework>
-    <Nullable>enable</Nullable>
-    <UseWindowsForms>true</UseWindowsForms>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="OpenTK" Version="4.7.1" />
-    <PackageReference Include="OpenTK.WinForms" Version="4.0.0-pre.6" />
-  </ItemGroup>
-
-</Project>
